requirements.txt
fastapi
uvicorn
scikit-learn
pandas
numpy
joblib
python-dotenv
streamlit
matplotlib
seaborn
requests